[REF] l10n_br_fiscal_subsequent_document:extracted
