+++ l10n_br_fiscal/__manifest__.py
@@ -10,7 +10,7 @@
     "maintainers": ["renatonlima", "rvalyi"],
     "website": "https://github.com/OCA/l10n-brazil",
     "development_status": "Production/Stable",
-    "version": "16.0.7.1.0",
+    "version": "16.0.8.0.0",
     "depends": [
         "product",
         "uom_alias",
@@ -97,8 +97,6 @@
         "views/document_line_view.xml",
         "views/document_fiscal_line_mixin_view.xml",
         "views/res_config_settings_view.xml",
-        "views/subsequent_operation_view.xml",
-        "views/subsequent_document_view.xml",
         "views/invalidate_number_view.xml",
         "views/city_taxation_code.xml",
         "views/operation_dashboard_view.xml",
@@ -119,7 +117,6 @@
         "demo/partner_demo.xml",
         "demo/fiscal_document_nfse_demo.xml",
         "demo/fiscal_operation_demo.xml",
-        "demo/subsequent_operation_demo.xml",
         "demo/res_users_demo.xml",
         "demo/icms_tax_definition_demo.xml",
         "demo/fiscal_document_demo.xml",
