+++ l10n_br_fiscal/models/__init__.py
@@ -51,7 +51,5 @@ from . import document_line
 from . import res_config_settings
 from . import res_country_state
 from . import operation_dashboard
-from . import subsequent_operation
-from . import subsequent_document
 from . import city_taxation_code
 from . import document_supplement
