+++ l10n_br_fiscal/tests/__init__.py
@@ -12,6 +12,5 @@ from . import (
     test_ncm,
     test_partner_profile,
     test_service_type,
-    test_subsequent_operation,
     test_operation,
 )
