+++ l10n_br_fiscal/views/document_view.xml
@@ -375,24 +375,6 @@
             </page>
             <page name="document_related" string="Document Related">
               <field name="document_related_ids" />
-              <separator colspan="4" string="Subsequents Documents" />
-              <field name="document_subsequent_ids">
-                  <tree editable="bottom">
-                    <field
-                                        name="subsequent_document_id"
-                                        context="{'fiscal_document_complete_name': True}"
-                                    />
-                    <field name="fiscal_operation_id" />
-                    <field name="operation_performed" />
-                    <field name="subsequent_operation_id" />
-                    <button
-                                        name="show_subsequent_document"
-                                        type="object"
-                                        string="Show Document"
-                                        attrs="{'invisible': [('subsequent_document_id', '=', False)]}"
-                                    />
-                </tree>
-              </field>
             </page>
             <page name="finance" string="Finance">
             </page>
