+++ l10n_br_fiscal/views/l10n_br_fiscal_action.xml
@@ -500,38 +500,6 @@
         </field>
     </record>
 
-    <!-- Subsequent Operation-->
-    <record id="subsequent_operation_action" model="ir.actions.act_window">
-        <field name="name">Subsequent Operation</field>
-        <field name="type">ir.actions.act_window</field>
-        <field name="res_model">l10n_br_fiscal.subsequent.operation</field>
-        <field name="view_mode">kanban,tree,form</field>
-        <field name="view_id" ref="subsequent_operation_tree" />
-        <field name="help" type="html">
-          <p class="o_view_nocontent_smiling_face">
-            Create a new Subsequent operation
-          </p><p>
-
-          </p>
-        </field>
-    </record>
-
-    <!-- Subsequent Document-->
-    <record id="subsequent_document_action" model="ir.actions.act_window">
-        <field name="name">Subsequent Document</field>
-        <field name="type">ir.actions.act_window</field>
-        <field name="res_model">l10n_br_fiscal.subsequent.document</field>
-        <field name="view_mode">kanban,tree,form</field>
-        <field name="view_id" ref="subsequent_document_tree" />
-        <field name="help" type="html">
-          <p class="o_view_nocontent_smiling_face">
-            Create a new Subsequent Document
-          </p><p>
-
-          </p>
-        </field>
-    </record>
-
     <!-- Comment -->
     <record id="comment_action" model="ir.actions.act_window">
         <field name="name">Comment</field>
