+++ l10n_br_fiscal/views/l10n_br_fiscal_menu.xml
@@ -515,15 +515,6 @@
         sequence="10"
     />
 
-    <!-- Subsequent Operation -->
-    <menuitem
-        id="subsequent_operation_menu"
-        action="subsequent_operation_action"
-        groups="l10n_br_fiscal.group_manager"
-        parent="operations_config_menu"
-        sequence="40"
-    />
-
     <!-- CFOP -->
     <menuitem
         id="cfop_menu"
@@ -576,15 +567,6 @@
         groups="l10n_br_fiscal.group_manager"
         parent="others_config_menu"
         sequence="50"
-    />
-
-    <!-- Subsequent Document -->
-    <menuitem
-        id="subsequent_document_menu"
-        action="subsequent_document_action"
-        groups="base.group_no_one"
-        parent="others_config_menu"
-        sequence="60"
     />
 
 </odoo>
