+++ l10n_br_fiscal_subsequent_document/models/__init__.py
@@ -1,0 +1,4 @@
+from . import subsequent_operation
+from . import subsequent_document
+from . import operation
+from . import document
