+++ l10n_br_fiscal_subsequent_document/readme/CONTRIBUTORS.md
@@ -1,0 +1,3 @@
+- [KMEE](https://www.kmee.com.br):
+  - Luis Felipe Mileo <mileo@kmee.com.br>
+  - Gabriel Cardoso de Faria <gabriel.cardoso@kmee.com.br>
