+++ l10n_br_fiscal_subsequent_document/readme/DESCRIPTION.md
@@ -1,0 +1,1 @@
+Conceito de operações subsequentes e documentos fiscais subsequentes.
