+++ l10n_br_fiscal_subsequent_document/tests/__init__.py
@@ -1,0 +1,1 @@
+from . import test_subsequent_operation
