+++ l10n_br_fiscal_subsequent_document/views/document_view.xml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+
+    <record id="document_form" model="ir.ui.view">
+        <field name="inherit_id" ref="l10n_br_fiscal.document_form" />
+        <field name="model">l10n_br_fiscal.document</field>
+        <field name="document_related_ids" position="after">
+            <separator colspan="4" string="Subsequents Documents" />
+            <field name="document_subsequent_ids">
+                <tree editable="bottom">
+                    <field
+                        name="subsequent_document_id"
+                        context="{'fiscal_document_complete_name': True}"
+                    />
+                    <field name="fiscal_operation_id" />
+                    <field name="operation_performed" />
+                    <field name="subsequent_operation_id" />
+                    <button
+                        name="show_subsequent_document"
+                        type="object"
+                        string="Show Document"
+                        attrs="{'invisible': [('subsequent_document_id', '=', False)]}"
+                    />
+                </tree>
+            </field>
+        </field>
+
+    </record>
+</odoo>
