+++ l10n_br_fiscal_subsequent_document/views/operation_view.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+
+    <record id="operation_form" model="ir.ui.view">
+        <field name="inherit_id" ref="l10n_br_fiscal.operation_form" />
+        <field name="model">l10n_br_fiscal.operation</field>
+        <page name="operation_document_type" position="after">
+            <page name="subsequent_operation" string="Operations">
+                <group>
+                    <field name="return_fiscal_operation_id" />
+                    <field name="inverse_fiscal_operation_id" />
+                </group>
+                <field name="operation_subsequent_ids" />
+            </page>
+        </page>
+    </record>
+
+</odoo>
