+++ l10n_br_fiscal_subsequent_document/views/subsequent_document_view.xml
