+++ setup/l10n_br_fiscal_subsequent_document/odoo/addons/l10n_br_fiscal_subsequent_document
@@ -1,0 +1,1 @@
+../../../../l10n_br_fiscal_subsequent_document
